pyproject.toml
[project]
name = "testhealer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=1.2.7",
    "langchain-anthropic>=1.3.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "rich>=14.3.1",
]
